use minio for dask.s3fs test too
